test(kt-devnet): refactor tests into memstore vs non-memstore
Added tests that can run against holesky devnet.
The memstore tests are kept separate because they require memstore-specific endpoits and functionality (such as causing 503 errors to be returned to test failover)
